Add scriptlets for migration away from systemd-timedated-ntp.target
F17 had systemd-timedated-ntp.target. The target is gone in F18.
We want to ensure that users who had their NTP services enabled
indirectly via the target in F17 will still have them enabled after
upgrading to F18.

We use the %pretrans step to record which NTP services to enable.
Doing it in %pre would be too late. %pretrans has to be in lua.
The actual enabling of the NTP services is done in %post.